+++ frontend/src/App.tsx
@@ -15,6 +15,7 @@ import { ENCRulesPage } from './pages/ENCRules';
 import { ConfigPuppetPage } from './pages/ConfigPuppet';
 import { ConfigPuppetDBPage } from './pages/ConfigPuppetDB';
 import { ConfigAppPage } from './pages/ConfigApp';
+import { UserManagerPage } from './pages/UserManager';
 
 function AppRoutes() {
   const { user, loading } = useAuth();
@@ -44,6 +45,9 @@ function AppRoutes() {
         <Route path="/enc/groups" element={<ENCGroupsPage />} />
         <Route path="/enc/classifications" element={<ENCClassificationsPage />} />
         <Route path="/enc/rules" element={<ENCRulesPage />} />
+
+        {/* Administration */}
+        <Route path="/users" element={<UserManagerPage />} />
 
         {/* Configuration */}
         <Route path="/config/puppet" element={<ConfigPuppetPage />} />